requirements.txt
pypdf
langchain
openai
pinecone-client
tiktoken
python-dotenv